Reorganize children components
Making it easier to identify the high level components.
