+++ web-app/src/com/lewismorgan/web/HeaderComponent.kt
@@ -1,11 +1,13 @@
 package com.lewismorgan.web
 
 import com.lewismorgan.web.bootstrap.dsl.container
+import com.lewismorgan.web.children.FontAwesomeStyleType
+import com.lewismorgan.web.children.fontAwesome
+import com.lewismorgan.web.children.gravatar
 import react.RBuilder
 import react.RComponent
 import react.RProps
 import react.RState
-import react.dom.div
 import react.dom.h1
 import react.dom.header
 import react.dom.i
