+++ web-app/src/com/lewismorgan/web/WebsiteNavbarComponent.kt
@@ -7,6 +7,9 @@ import com.lewismorgan.web.bootstrap.nav.navbar
 import com.lewismorgan.web.bootstrap.nav.navbarBrand
 import com.lewismorgan.web.bootstrap.nav.navbarCollapse
 import com.lewismorgan.web.bootstrap.nav.navbarToggler
+import com.lewismorgan.web.children.FontAwesomeStyleType
+import com.lewismorgan.web.children.floatingItem
+import com.lewismorgan.web.children.fontAwesome
 import com.lewismorgan.web.misc.getSmoothScrollingHandler
 import com.lewismorgan.web.misc.navHashLink
 import org.w3c.dom.events.Event
@@ -16,6 +19,7 @@ import react.RProps
 import react.RState
 import react.dom.i
 import react.setState
+import kotlin.browser.window
 
 interface WebsiteNavbarProps : RProps {
   var collapsedMenuShown: Boolean
@@ -23,15 +27,38 @@ interface WebsiteNavbarProps : RProps {
 
 interface WebsiteNavbarState : RState {
   var collapsed: Boolean
+  var scrollMinimized: Boolean
 }
 
 class WebsiteNavbarComponent(props: WebsiteNavbarProps) : RComponent<WebsiteNavbarProps, WebsiteNavbarState>(props) {
   override fun WebsiteNavbarState.init() {
     collapsed = true
+    scrollMinimized = false
   }
 
   override fun WebsiteNavbarState.init(props: WebsiteNavbarProps) {
     collapsed = props.collapsedMenuShown
+    scrollMinimized = false
+  }
+
+  override fun componentDidMount() {
+    window.addEventListener("scroll", onScroll())
+  }
+
+  override fun componentWillUnmount() {
+    window.removeEventListener("scroll", onScroll())
+  }
+
+  private fun onScroll(): (Event) -> Unit = {
+    if (window.scrollY > 100) {
+      setState {
+        scrollMinimized = true
+      }
+    } else {
+      setState {
+        scrollMinimized = false
+      }
+    }
   }
 
   private fun onClickToggler(): (Event) -> Unit = {
@@ -47,7 +74,7 @@ class WebsiteNavbarComponent(props: WebsiteNavbarProps) : RComponent<WebsiteNavb
   }
 
   override fun RBuilder.render() {
-    navbar("navbar-expand-lg navbar-dark fixed-top bg-dark", true) {
+    navbar("navbar-expand-lg navbar-dark fixed-top bg-dark ${if (state.scrollMinimized) "navbar-shrink" else ""}") {
       container {
         navbarBrand("#") {
           fontAwesome("map-marked-alt", FontAwesomeStyleType.SOLID)
