+++ web-app/src/com/lewismorgan/web/children/FloatingItemComponent.kt
@@ -1,4 +1,4 @@
-package com.lewismorgan.web
+package com.lewismorgan.web.children
 
 import com.lewismorgan.web.bootstrap.dsl.container
 import com.lewismorgan.web.misc.getScrollTop
