+++ web-app/src/com/lewismorgan/web/children/GravatarProfileComponent.kt
@@ -1,4 +1,4 @@
-package com.lewismorgan.web
+package com.lewismorgan.web.children
 
 
 import kotlinx.html.classes
